Implement multiton pattern in plugin manager

+++ tests/Bootstrap.php
@@ -4,6 +4,9 @@ include __DIR__ . "/../vendor/autoload.php";
 
 function delTree($dir)
 {
+    if (!file_exists($dir)) {
+        return;
+    }
     $files = array_diff(scandir($dir), array('.', '..'));
     foreach ($files as $file) {
         (is_dir("$dir/$file")) ? delTree("$dir/$file") : unlink("$dir/$file");
